requirements.txt
openai>=1.3.0
requests>=2.28.0
python-dotenv>=0.21.0
rich>=12.0.0 